fix some types

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/Constants.java
@@ -70,5 +70,5 @@ public class Constants {
   public static final int EPOCHS_PER_VALIDATOR_REGISTRATION_SUBMISSION = 1;
 
   // ZkChain prototype
-  public static final int MAX_EXECUTION_PROOF_SUBNETS = 8;
+  public static final long MAX_EXECUTION_PROOF_SUBNETS = 8;
 }
