+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/gossip/subnets/PeerSubnetSubscriptions.java
@@ -95,7 +95,7 @@ public class PeerSubnetSubscriptions {
         builder(
                 currentSchemaDefinitions,
                 SszBitvectorSchema.create(dataColumnSidecarSubnetCount),
-                SszBitvectorSchema.create(MAX_EXECUTION_PROOF_SUBNETS.intValue()))
+                SszBitvectorSchema.create(MAX_EXECUTION_PROOF_SUBNETS))
             .targetSubnetSubscriberCount(targetSubnetSubscriberCount)
             .nodeIdToDataColumnSidecarSubnetsCalculator(nodeIdToDataColumnSidecarSubnetsCalculator)
             .attestationSubnetSubscriptions(
